requirements.txt
# Core Dependencies - Essential for bot functionality
python-telegram-bot>=20.7
python-dotenv>=1.0.0
psutil>=5.9.5

# Basic utilities (lightweight)
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3

# Optional AI features (uncomment if needed)
# openai>=1.3.7
# anthropic>=0.8.0

# Optional VPS management (uncomment if needed)
# paramiko>=3.3.1

# Optional Docker management (uncomment if needed)
# docker>=6.1.3

# Optional database (uncomment if needed)
# asyncpg>=0.29.0
